+++ R/ga_data.R
@@ -2,6 +2,54 @@ version_aw <- function(){
   "v1alpha"
 }
 
+#' Extract metric aggregations from a \link{ga_data} result
+#' 
+#' Metric aggregations are available in each \link{ga_data} result.  This function lets you easily access the data.frames
+#' 
+#' @param df A data.frame result from \link{ga_data}
+#' @param type totals, maximums, minimums, counts (if available) or all
+#' 
+#' @export
+#' @examples 
+#' 
+#' \dontrun{
+#' #' # send up to 4 date ranges
+#' multi_date <- ga_data(
+#'   206670707,
+#'   metrics = c("activeUsers","sessions"),
+#'   dimensions = c("date","city","dayOfWeek"),
+#'   date_range = c("2020-03-31", "2020-04-27", "2020-04-30", "2020-05-27"),
+#'   dimensionFilter = ga_data_filter("city"=="Copenhagen"),
+#'   limit = 100
+#'   )
+#'
+#' # metric aggregations for each date range
+#' ga_data_aggregations(multi_date, type = "all")
+#' 
+#' # specify type
+#' ga_data_aggregations(multi_date, type = "maximums")
+#' 
+#' }
+ga_data_aggregations <- function(df, 
+                                 type = c("all","totals",
+                                          "maximums","minimums",
+                                          "count")){
+  type <- match.arg(type)
+  if(is.null(attr(df, "metricAggregations"))){
+    stop("No aggregations found.  Is the data.frame from ga_data()?", 
+         call. = FALSE)
+  }
+  
+  ma <- attr(df, "metricAggregations")
+  
+  if(type == "all"){
+    return(ma)
+  }
+  
+  ma[[type]]
+  
+}
+
 
 #' Google Analytics Data for GA4 (App+Web)
 #'
@@ -19,6 +67,8 @@ version_aw <- function(){
 #' @importFrom googleAuthR gar_api_generator
 #' @import assertthat
 #' @family GA4 functions
+#' 
+#' @return A data.frame tibble, including attributes metadata, metricAggregations and rowCount.  Use \link{ga_data_aggregations} to extract the data.frames of metricAggregations
 #' @export
 #' @examples 
 #' 
@@ -63,6 +113,8 @@ version_aw <- function(){
 #'   limit = 100,
 #'   realtime = TRUE)
 #' 
+#' # extract meta data from the table
+#' ga_data_aggregations(realtime)
 #' }
 ga_data <- function(propertyId,
                     metrics,
@@ -71,7 +123,6 @@ ga_data <- function(propertyId,
                     dimensionFilter = NULL,
                     dimensionDelimiter = "/",
                     metricFilter = NULL,
-                    # metricAggregations = NULL,
                     orderBys = NULL,
                     limit = 100,
                     realtime=FALSE) {
@@ -80,15 +131,8 @@ ga_data <- function(propertyId,
   dimensionFilter <- as_filterExpression(dimensionFilter)
   metricFilter    <- as_filterExpression(metricFilter)
  
-  
-  # if(!is.null(metricAggregations)){
-  #   assert_that(all(metricAggregations %in% c("TOTAL",
-  #                                              "MINIMUM",
-  #                                              "MAXIMUM",
-  #                                              "COUNT")))
-  # }
- 
-  metricAggregations <- NULL
+  # we always get these 3 - COUNT is not available unless pivot?
+  metricAggregations <- c("TOTAL","MAXIMUM","MINIMUM")
   
   dims <- gaw_dimension(dimensions, delimiter = dimensionDelimiter)
   mets <- gaw_metric(metrics)
@@ -156,7 +200,8 @@ ga_data <- function(propertyId,
 #' @noRd
 ga_aw_realtime <- function(property, requestObj){
   
-  url <- sprintf("https://analyticsdata.googleapis.com/%s/properties/%s:runRealtimeReport",
+  url <- 
+    sprintf("https://analyticsdata.googleapis.com/%s/properties/%s:runRealtimeReport",
                  version_aw(), property)
   # analyticsdata.runRealtimeReport
   f <- gar_api_generator(url, "POST", 
@@ -253,6 +298,28 @@ row_types <- function(res, met_names){
 
 #' @noRd
 #' @importFrom dplyr bind_cols bind_rows across mutate
+parse_aggregations <- function(agg, dim_names, met_names){
+  
+  dds <- lapply(agg$dimensionValues, function(x) setNames(x$value, dim_names))
+  mms <- lapply(agg$metricValues, function(x) setNames(x$value, met_names))
+  
+  dds <- bind_rows(dds)
+  mms <- bind_rows(mms)
+  
+  # bind_cols returns 0rows if first df has 0
+  if(nrow(dds) == 0){
+    res <- mms
+  } else {
+    res <- bind_cols(dds, mms)
+  }
+  
+  res <- row_types(res, met_names)
+  
+  res
+}
+
+#' @noRd
+#' @importFrom dplyr bind_cols bind_rows across mutate
 parse_rows <- function(o, dim_names, met_names){
 
   quota_messages(o)
@@ -265,14 +332,31 @@ parse_rows <- function(o, dim_names, met_names){
   
   # bind_cols returns 0rows if first df has 0
   if(nrow(dds) == 0){
-    res <- bind_rows(mms)
+    res <- mms
   } else {
     res <- bind_cols(dds, mms)
   }
   
   res <- row_types(res, met_names = met_names)
-
+  
+  dl <- attr(res, "metadata")[["dataLossFromOtherRow"]]
+  if(!is.null(dl) && dl){
+    myMessage("Warning: some buckets of dimension combinations are rolled into '(other)' row. This can happen for high cardinality reports.", level = 3)
+  }
+  
   attr(res, "metadata") <- o$metadata
+  attr(res, "metricAggregations") <- list(
+    totals =   parse_aggregations(o$totals, dim_names, met_names),
+    maximums = parse_aggregations(o$maximums, dim_names, met_names),
+    minimums = parse_aggregations(o$minimums, dim_names, met_names)
+  )
+  attr(res, "rowCount") <- o$rowCount
+  
+  ## remove dateRange column if only one unique value
+  if(!is.null(res$dateRange) &&
+     all(unique(res$dateRange) == "date_range_0")){
+    res$dateRange <- NULL
+  }
   
   res